Initial work on fortplan plugin, including separating out code that needs to be shared with buildingplan
