+++ library/include/Hooks.h
@@ -47,6 +47,7 @@ namespace SDL
 // these functions are here because they call into DFHack::Core and therefore need to
 // be declared as friend functions/known
 #ifdef _DARWIN
+#include "modules/Graphic.h"
 DFhackCExport int DFH_SDL_NumJoysticks(void);
 DFhackCExport void DFH_SDL_Quit(void);
 DFhackCExport int DFH_SDL_PollEvent(SDL::Event* event);
