+++ plugins/CMakeLists.txt
@@ -93,7 +93,7 @@ if (BUILD_SUPPORTED)
     DFHACK_PLUGIN(autotrade autotrade.cpp)
     DFHACK_PLUGIN(burrows burrows.cpp LINK_LIBRARIES lua)
     DFHACK_PLUGIN(building-hacks building-hacks.cpp LINK_LIBRARIES lua)
-    DFHACK_PLUGIN(buildingplan buildingplan.cpp)
+    DFHACK_PLUGIN(buildingplan buildingplan-lib.cpp buildingplan.cpp)
     DFHACK_PLUGIN(catsplosion catsplosion.cpp)
     DFHACK_PLUGIN(changeitem changeitem.cpp)
     DFHACK_PLUGIN(changelayer changelayer.cpp)
@@ -122,6 +122,7 @@ if (BUILD_SUPPORTED)
     DFHACK_PLUGIN(flows flows.cpp)
     DFHACK_PLUGIN(follow follow.cpp)
     DFHACK_PLUGIN(forceequip forceequip.cpp)
+    DFHACK_PLUGIN(fortplan buildingplan-lib.cpp fortplan.cpp)
     DFHACK_PLUGIN(getplants getplants.cpp)
     DFHACK_PLUGIN(hotkeys hotkeys.cpp)
     DFHACK_PLUGIN(infiniteSky infiniteSky.cpp)
